# src/main/java/com/rajharit/rajharitsprings/dtos/StockMovementDto.java
package com.rajharit.rajharitsprings.dtos;

import lombok.Data;

import java.time.LocalDateTime;

@Data
public class StockMovementDto {
    private String movementType;
    private double quantity;
    private String unit;
    private LocalDateTime movementDate;
}